Layout refactor and standardization
