+++ src/components/index.ts
@@ -1,0 +1,4 @@
+import BlockLink from "./BlockLink";
+import Layout from "./Layout";
+
+export { BlockLink, Layout };
